test: reconfigure between map & surface role tests
The brief period between a window being mapped and obtaining a surface
role did not respect pop-up reconfiguration. This adds a test for that,
along with a test that the toplevel's position cannot be reconfigured
after being mapped.

PR #281 contains the fixes, and therefore should pass these tests. I
decided to write them since the PR author struggled with them and I
realized the implementation would be somewhat involved.